HTBHF-2045 move non apply route handlers to correct directory
